requirements.txt
# Core dependencies
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Telegram bot
python-telegram-bot>=20.0

# Technical indicators
ta>=0.10.0
# talib-binary>=0.4.26  # Optional, install separately if needed

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Utilities
aiohttp>=3.8.0
requests>=2.28.0

# Logging and monitoring
colorlog>=6.7.0